make SofaBenchmarks scenes a standalone project
