+++ SofaBenchmarkScenes/src/Main.cpp
@@ -1,0 +1,3 @@
+#include <benchmark/benchmark.h>
+
+BENCHMARK_MAIN();
